+++ .pre-commit-config.yaml
@@ -51,7 +51,7 @@ repos:
   hooks:
     - id: generate-init
       name: "Check generated __init__.py"
-      entry: ./src/globus_sdk/__generate_init.py
+      entry: ./src/globus_sdk/_generate_init.py
       language: script
       always_run: true
       pass_filenames: false
