+++ src/globus_sdk/services/gcs/data/storage_gateway.py
@@ -2,7 +2,7 @@ import abc
 from typing import Any, Dict, Iterable, Optional, Tuple, Union
 
 from globus_sdk import utils
-from globus_sdk.types import UUIDLike
+from globus_sdk._types import UUIDLike
 
 from ._common import ensure_datatype
